Successful compilation! Now to see how broken it all is
