+++ src/app/data/items/CategoryBoostingItem.ts
@@ -1,6 +1,6 @@
 import { MoveData } from "@/app/damagecalc/components/MoveCard";
 import { DamageMultipliers } from "@/app/damagecalc/damageCalc";
-import { LoadedItem } from "@/preload/loadTectonicRepoData";
+import { LoadedItem } from "@/preload/loadedDataClasses";
 import { Item } from "../tectonic/Item";
 import { PartyPokemon } from "../types/PartyPokemon";
