+++ src/app/data/items/FlatDamageBoostItem.ts
@@ -1,5 +1,5 @@
 import { DamageMultipliers } from "@/app/damagecalc/damageCalc";
-import { LoadedItem } from "@/preload/loadTectonicRepoData";
+import { LoadedItem } from "@/preload/loadedDataClasses";
 import { Item } from "../tectonic/Item";
 
 const itemBoosts: Record<string, number> = {
