+++ src/app/data/items/TypeBoostingItem.ts
@@ -1,6 +1,6 @@
 import { MoveData } from "@/app/damagecalc/components/MoveCard";
 import { DamageMultipliers } from "@/app/damagecalc/damageCalc";
-import { LoadedItem } from "@/preload/loadTectonicRepoData";
+import { LoadedItem } from "@/preload/loadedDataClasses";
 import { BattleState } from "../battleState";
 import { Item } from "../tectonic/Item";
 import { PokemonType } from "../tectonic/PokemonType";
