+++ src/app/data/moves/ConditionalAutoBoostMove.ts
@@ -1,4 +1,4 @@
-import { LoadedMove } from "@/preload/loadTectonicRepoData";
+import { LoadedMove } from "@/preload/loadedDataClasses";
 import { BattleState } from "../battleState";
 import { StatusEffect } from "../conditions";
 import { Move } from "../tectonic/Move";
