+++ src/app/data/moves/DifferentAttackStatMove.ts
@@ -1,4 +1,4 @@
-import { LoadedMove } from "@/preload/loadTectonicRepoData";
+import { LoadedMove } from "@/preload/loadedDataClasses";
 import { Move } from "../tectonic/Move";
 import { Stat } from "../tectonic/Pokemon";
